fixes #2899 : Copied list, position is not updated correctly in other user browser issue fixed.

+++ client/js/views/footer_view.js
@@ -1666,7 +1666,7 @@ App.FooterView = Backbone.View.extend({
                                             new_list.set('board_id', parseInt(activity.attributes.list.board_id));
                                             new_list.set('lists_cards', []);
                                             new_list.set('is_archived', 0);
-                                            new_list.set('position', parseInt(activity.attributes.list.position));
+                                            new_list.set('position', parseFloat(activity.attributes.list.position));
                                             self.board.lists.add(new_list);
                                             if (self.board.attributes.lists === null) {
                                                 self.board.attributes.lists = [];
